re-add Granularity types
