+++ tests/helpers/test_monitor_helpers.py
@@ -6,6 +6,7 @@ from whylabs_toolkit.helpers.monitor_helpers import (
     get_monitor_config
 )
 from whylabs_toolkit.helpers.utils import get_monitor_api
+from whylabs_toolkit.utils.granularity import Granularity
 
 
 ORG_ID = os.environ["ORG_ID"]
@@ -91,7 +92,7 @@ class TestDeleteMonitor(BaseTestMonitor):
 
     def test_get_granularity(self) -> None:
         granularity = get_model_granularity(org_id=ORG_ID, dataset_id=DATASET_ID)
-        assert granularity == "monthly"
+        assert granularity == Granularity.monthly
 
 
     def test_delete_monitor(self) -> None:
