+++ whylabs_toolkit/monitor/__init__.py
@@ -1,5 +1,5 @@
 from .manager import MonitorSetup, MonitorManager
-from .models import *
+
 
 ALL = [
     MonitorManager,
