+++ whylabs_toolkit/utils/granularity.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+
+class Granularity(str, Enum):
+    """Supported granularity."""
+
+    hourly = "hourly"
+    daily = "daily"
+    weekly = "weekly"
+    monthly = "monthly"
